👌 IMPROVE: Seperate signup body and page
